Correct display of time in hours and minutes.
This change makes the preparation and cooking time being correctly displayed online.
The times which are stored in minutes are now correctly reported as hours+minutes.

Also not a small change in the definition of `TimesInMinsAdapted`, to ensure that
the time are reported as integers instead od doubles on the display.

+++ ui/app/com/gu/recipeasy/controllers/Application.scala
@@ -163,10 +163,10 @@ object Application {
       )(DetailedIngredientsList.apply)(DetailedIngredientsList.unapply)),
       "credit" -> optional(text(maxLength = 200)),
       "times" -> mapping(
-        "preparationHours" -> optional(of[Double]),
-        "preparationMinutes" -> optional(of[Double]),
-        "cookingHours" -> optional(of[Double]),
-        "cookingMinutes" -> optional(of[Double])
+        "preparationHours" -> optional(of[Int]),
+        "preparationMinutes" -> optional(of[Int]),
+        "cookingHours" -> optional(of[Int]),
+        "cookingMinutes" -> optional(of[Int])
       )(TimesInMinsAdapted.apply)(TimesInMinsAdapted.unapply),
       "steps" -> seq(text),
       "tags" -> mapping(
